+++ Sources/CometDClient/CometdClientDelegate.swift
@@ -14,13 +14,13 @@ public protocol CometdClientDelegate: class {
   func didReceivePong(from client: CometdClientContract)
   func didConnected(from client: CometdClientContract)
   func handshakeDidSucceeded(dictionary: NSDictionary, from client: CometdClientContract)
-  func handshakeDidFailed(from client: CometdClientContract)
+  func handshakeDidFailed(error: Error, from client: CometdClientContract)
   func didDisconnected(error: Error?, from client: CometdClientContract)
   func didAdvisedToReconnect(from client: CometdClientContract)
-  func didFailConnection(error: Error?, from client: CometdClientContract)
+  func didLostConnection(error: Error, from client: CometdClientContract)
   func didSubscribeToChannel(channel: String, from client: CometdClientContract)
   func didUnsubscribeFromChannel(channel: String, from client: CometdClientContract)
-  func subscriptionFailedWithError(error: SubscriptionError, from client: CometdClientContract)
+  func subscriptionFailedWithError(error: Error, from client: CometdClientContract)
   func didWriteError(error: Error, from client: CometdClientContract)
 }
 
@@ -29,12 +29,12 @@ public extension CometdClientDelegate {
   func didReceivePong(from client: CometdClientContract) { }
   func didConnected(from client: CometdClientContract) { }
   func handshakeDidSucceeded(dictionary: NSDictionary, from client: CometdClientContract) { }
-  func handshakeDidFailed(from client: CometdClientContract) { }
+  func handshakeDidFailed(error: Error, from client: CometdClientContract) { }
   func didDisconnected(error: Error?, from client: CometdClientContract) { }
   func didAdvisedToReconnect(from client: CometdClientContract) { }
-  func didFailConnection(error: Error?, from client: CometdClientContract) { }
+  func didLostConnection(error: Error, from client: CometdClientContract) { }
   func didSubscribeToChannel(channel: String, from client: CometdClientContract) { }
   func didUnsubscribeFromChannel(channel: String, from client: CometdClientContract) { }
-  func subscriptionFailedWithError(error: SubscriptionError, from client: CometdClientContract) { }
+  func subscriptionFailedWithError(error: Error, from client: CometdClientContract) { }
   func didWriteError(error: Error, from client: CometdClientContract) { }
 }