+++ Sources/CometDClient/Transport/Transport.swift
@@ -20,8 +20,8 @@ public protocol Transport {
 
 public protocol TransportDelegate: class {
   func didConnect()
-  func didFailConnection(_ error: Error?)
-  func didDisconnect(_ error: Error?)
+  func didLostConnection(_ error: Error)
+  func didDisconnect(_ error: Error)
   func didWriteError(_ error: Error?)
   func didReceiveMessage(_ text: String)
   func didReceivePong()