+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # CHANGELOG
 All notable changes to this project will be documented in this file.
 
+### [1.1.0](https://github.com/Insurlytech/CometDClient-iOS/1.1.0)
+
+* Errors - Create recorder for sharing important errors  
+* Errors - Refact error to be more precise on the failure  
+
 ### [1.0.0](https://github.com/Insurlytech/CometDClient-iOS/1.0.0)
  
  * First release of CometD Client for iOS